Got rid of ImageRequestBuilder_v3 interface
